feat(flask_app): 🛠️ Add tests
to run all tests write
```
pytest
```
in terminal
