+++ flask_app/resources/user.py
@@ -28,12 +28,19 @@ class UserResource(Resource):
     # Create a new user
     def post(self):
         args = user_parser.parse_args()
+
+        # Check if a user with the same email already exists
+        existing_user = Users.query.filter_by(email=args['email']).first()
+        if existing_user:
+            abort(400, message='User with the same email already exists')
+
         new_user = Users(
             name=args['name'],
             email=args['email'],
             phone=args['phone'],
             password=generate_password_hash(args['password'])
         )
+
         db.session.add(new_user)
         db.session.commit()
         return {'message': 'User created'}, 201
